+++ .gitignore
@@ -3,3 +3,4 @@
 .DS_Store
 .vscode/settings.json
 *.svg
+*.png
